refactor ViewHolder.kt into /util
