+++ app/src/androidTest/java/com/github/fribourgsdp/radio/PlaylistRecyclerViewTest.kt
@@ -12,19 +12,15 @@ import androidx.test.espresso.contrib.RecyclerViewActions
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.mockimplementations.MockUserProfileActivity
 import com.github.fribourgsdp.radio.mockimplementations.playListName
 import com.github.fribourgsdp.radio.mockimplementations.songName
+import com.github.fribourgsdp.radio.util.ViewHolder
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.auth.FirebaseAuth
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import org.junit.After
 import org.junit.Before
-import org.junit.Rule
-import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.concurrent.TimeUnit
