+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/PlaylistSongsFragment.kt
@@ -12,6 +12,7 @@ import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.util.MyFragment
+import com.github.fribourgsdp.radio.util.OnClickListener
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
