+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/UserPlaylistsFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.fribourgsdp.radio.util.MyFragment
-import com.github.fribourgsdp.radio.OnClickListener
+import com.github.fribourgsdp.radio.util.OnClickListener
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.User
